requirements.txt
requests
pillow
python-dotenv
openai
Flask
Flask-SocketIO
pymongo
markdown
pyttsx3
